# apps/books-service/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing books in Shelfshare.",
        "title": "Shelfshare Books API",
        "contact": {
            "name": "Sina Niyavarzi",
            "email": "sinaniya@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
        "/authors": {
            "get": {
                "description": "Get a list of all authors",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["authors"],
                "summary": "List authors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handler.AuthorResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new author with name and optional bio",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["authors"],
                "summary": "Create an author",
                "parameters": [
                    {
                        "description": "Author to create",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.CreateAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handler.AuthorResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/authors/{id}": {
            "get": {
                "description": "Get a single author by its ID",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["authors"],
                "summary": "Get author by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AuthorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an author by ID",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["authors"],
                "summary": "Delete an author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid ID",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Partially update an existing author",
                "consumes": ["application/json"],
                "produces": ["application/json"],
                "tags": ["authors"],
                "summary": "Update an author",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Author ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Author fields to update",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.UpdateAuthorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AuthorResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid ID or validation error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Login required",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Author not found",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/validation.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Author": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string"
                },
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.BookSummary"
                    }
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-11-24"
                }
            }
        },
        "handler.AuthorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/handler.Author"
                }
            }
        },
...
        "handler.BookSummary": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string",
                    "example": "2025-11-24"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-11-24"
                }
            }
        },
        "handler.CreateAuthorRequest": {
            "type": "object",
            "required": ["name"],
            "properties": {
                "bio": {
                    "type": "string",
                    "maxLength": 2000
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "handler.UpdateAuthorRequest": {
            "type": "object",
            "properties": {
                "bio": {
                    "type": "string",
                    "maxLength": 2000
                },
                "name": {
                    "type": "string",
                    "minLength": 1
                }
            }
        },
...
        "validation.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/validation.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "validation.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "rule": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Supply the auth-service JWT as: Bearer \u003ctoken\u003e. Required for all mutations.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}